LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_entry_type-expected.txt

CONSOLE MESSAGE: line 25: ReferenceError: Can't find variable: test_equals
Description

This test validates the user timing entry type, PerformanceMark and PerformanceMeasure.


Harness Error (FAIL), message = ReferenceError: Can't find variable: test_equals
